+++ packages/modules/kit-feedback/src/lib/components/Vote.svelte
