+++ packages/modules/kit-feedback/src/lib/components/comments/COMMENT_List.svelte
@@ -2,7 +2,7 @@
 	import { config } from '$lib/config';
 	import { resolveTheme, theme } from '$lib/theme';
 	import type { Comment as TComment } from '$lib/types';
-	import Comment from './Comment.svelte';
+	import CommentDetail from '$lib/components/comments/COMMENT_Detail.svelte';
 
 	export let comments: TComment[];
 </script>
@@ -10,7 +10,7 @@
 <div class={resolveTheme($theme, 'comments-list')}>
 	{#each comments as comment}
 		{#if $config.dev || comment.public}
-			<Comment {comment} />
+			<CommentDetail {comment} />
 		{/if}
 	{/each}
 </div>
