+++ packages/modules/kit-feedback/src/lib/components/milestones/MILESTONE_List.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { MilestonePreviewFragment } from '$lib/graphql/_kitql/graphqlTypes';
-	import MilestonePreview from '$lib/components/milestones/MilestonePreview.svelte';
+	import MilestonePreview from '$lib/components/milestones/MILESTONE_Preview.svelte';
 
 	export let milestones: MilestonePreviewFragment[];
 </script>
