+++ packages/modules/kit-feedback/src/lib/pages/Milestones.svelte
@@ -3,7 +3,7 @@
 	import { config } from '$lib/config';
 	import { KQL_Milestones } from '$lib/graphql/_kitql/graphqlStores';
 	import Icon from '@iconify/svelte';
-	import MilestonesList from '$lib/components/milestones/MilestonesList.svelte';
+	import MilestoneList from '$lib/components/milestones/MILESTONE_List.svelte';
 
 	export let take: number = 25;
 
@@ -21,5 +21,5 @@
 {#if $KQL_Milestones.status === 'LOADING'}
 	<Icon icon="eos-icons:loading" />
 {:else}
-	<MilestonesList milestones={$KQL_Milestones.data?.repository?.milestones?.nodes} />
+	<MilestoneList milestones={$KQL_Milestones.data?.repository?.milestones?.nodes} />
 {/if}
